chore: rename useSend hook
